# BACKEND/02-back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Click2Book API",
    "description": "Online Ticket Reservation System — REST API documentation for Click2Book (IIIT Sri City FSD 2024–2028, Team 13)",
    "version": "1.0.0",
    "contact": {
      "name": "Team 13 — Click2Book",
      "url": "https://github.com/IIIT-Sricity-FSD-2024-2028/13_Click2Book"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
...
  "paths": {
...
    "/schedule/{id}": {
      "get": {
        "tags": ["Schedule"],
        "summary": "Get schedule by ID",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Schedule details" },
          "404": { "description": "Schedule not found" }
        }
      },
      "put": {
        "tags": ["Schedule"],
        "summary": "Update schedule",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateScheduleDto" } } }
        },
        "responses": {
          "200": { "description": "Schedule updated" },
          "404": { "description": "Not found" }
        }
      },
      "delete": {
        "tags": ["Schedule"],
        "summary": "Delete schedule",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Schedule deleted" },
          "404": { "description": "Not found" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "CreateScheduleDto": {
        "type": "object",
        "required": ["routeId", "vehicleId", "departureTime", "arrivalTime", "fare"],
        "properties": {
          "routeId": { "type": "string" },
          "vehicleId": { "type": "string" },
          "departureTime": { "type": "string", "format": "date-time" },
          "arrivalTime": { "type": "string", "format": "date-time" },
          "fare": { "type": "number" }
        }
      },
...
    }
  }
}